Update README.md
This project is focused on building a machine learning model to recognize handwritten digits, like those found in the famous MNIST dataset. The MNIST dataset contains thousands of images of digits (0-9), each labeled with the correct digit.

The main goal of the project is to train a model that can accurately predict which digit is shown in an image. Here's how the project works:

Data Preparation: The images from the MNIST dataset are loaded and preprocessed to be in a format suitable for training a machine learning model.

Model Training: The project uses the Random Forest Classifier to train on the dataset and learn how to recognize the digits. This algorithm is particularly effective for classification tasks like this one, where it builds multiple decision trees and merges them together to improve the accuracy of predictions.

Model Evaluation: The performance of these models is tested using accuracy and other metrics to determine how well the models are able to predict digits in new, unseen images.

Prediction: Once trained, the model can be used to predict the digit in any new handwritten image.